requirements.txt
gunicorn
flask
pyrogram
tgcrypto
python-dotenv
waitress